renamed new component to CredTimline to ExplorerTimeline
